fix: better error messages during the runs
This commit improves the logging for this GitHub Action
so developer can understand most of the thing from the console
output without a need to look into the HTML report.

This commit also removes couple of simple functions and embeds
their functionality into the main function without losing
readability.

The HTML report will be anyways downloaded and stored for Splunk
repositories.